Migrate to org-mode 9.5
This is a major update of org-mode, which required some changes
to init.el.

Additionally, I have waffled, once again, on org-adapt-indentation
and turned it back off. But this comes with the addition of using
the null character (?\00) as the leading bullet in org-superstar,
which means I finally like my org-mode headlines.

I may yet make level-1 and level-2 headlines larger, though.